Extract cli specific parameters from config to cli_configuration file (#9)
