+++ src/config.py
@@ -1,6 +1,5 @@
 import configparser
 from functools import cache
-from typing import Optional
 
 from src.custom_types import WatermarkRelativeSize
 from src.ffmpeg_utils_mixin import FFmpegUtilsMixin
@@ -12,57 +11,11 @@ class _ConfigManager(MediaUtilsMixin, FFmpegUtilsMixin):
     WATERMARK_POSITIONS = ["NE", "NC", "NW", "SE", "SC", "SW", "C", "CE", "CW"]
 
     def __init__(self) -> None:
-        self._keep_output_tree = None
-        self._overwrite: Optional[bool] = None
-        self._output_file_prefix: Optional[str] = None
-        self._watermark_file_path: Optional[str] = None
-        self._output_dir_path: Optional[str] = None
-
         self.config = configparser.ConfigParser()
         self.config.read(_ConfigManager.configuration_file_path)
 
     def get(self, section: str, key: str) -> str:
         return self.config.get(section, key)
-
-    @property
-    def keep_output_tree(self) -> Optional[bool]:
-        return self._keep_output_tree
-
-    @keep_output_tree.setter
-    def keep_output_tree(self, keep_output_tree: bool) -> None:
-        self._keep_output_tree = keep_output_tree
-
-    @property
-    def overwrite(self) -> Optional[bool]:
-        return self._overwrite
-
-    @overwrite.setter
-    def overwrite(self, overwrite: bool) -> None:
-        self._overwrite = overwrite
-
-    @property
-    def output_file_prefix(self) -> Optional[str]:
-        return self._output_file_prefix
-
-    @output_file_prefix.setter
-    def output_file_prefix(self, prefix: str) -> None:
-        self._output_file_prefix = prefix
-
-    @property
-    def output_dir_path(self) -> Optional[str]:
-        return self._output_dir_path
-
-    @output_dir_path.setter
-    def output_dir_path(self, dir_path: str) -> None:
-        self._output_dir_path = dir_path
-
-    @property
-    def watermark_file_path(self) -> Optional[str]:
-        return self._watermark_file_path
-
-    @watermark_file_path.setter
-    def watermark_file_path(self, file_path: str) -> None:
-        self._watermark_file_path = file_path
 
     @property
     @cache
